nyitottagy/requirements.txt
pandas
scipy.spatial.distance
